+++ app/services/bluControl.py
@@ -57,6 +57,7 @@ class BluetoothCon():
         return result
 
     def connect_device(self, mac):
+        print(f"MAC connect request: {mac}")
         stat = False
         if self.platform == "android":
             try:
@@ -79,7 +80,7 @@ class BluetoothCon():
                 print(f"Error while conencting to ESP32: {e}")
         return stat
 
-    def send_cmd(self, cmd):
+    def send_cmd(self, cmd:str, *args):
         stat = False
         if self.platform == "android" and self.connect_ok:
             try: