+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Changes
 
+### 0.0.3
+- Now app will use a service, so that app may not need to stay on foreground.
+- Now automatic nagivation start threashold is 60m (earlier 50m).
+- Will show an error if Bluetooth connection fails.
+- Updated correct demo & documentation link.
+- Adding android api 36 (android 16) support.
+- Some bug fixes.
+
 ### 0.0.2
 - Introducing the `Bluetooth` connectivity for ESP32.
 - Adding the navigation app API logic.